refactor: move logic to useGetTodo, delete unused files
